+++ test/plugin/scenarios/servicecomb-0.x-scenario/config/expectedData.yaml
@@ -16,18 +16,16 @@
 
 registryItems:
   applications:
-    - {servicecomb-provider: nq 0}
-    - {servicecomb-consumer: nq 0}
+    - {servicecomb-0.x-scenario: 2}
   instances:
-    - {servicecomb-provider: 1}
-    - {servicecomb-consumer: 1}
+    - {servicecomb-0.x-scenario: 1}
   operationNames:
-    - servicecomb-consumer: [codefirstClient.codeFirstSpringmvcHello.say, /servicecomb/case]
-    - servicecomb-provider: [codefirst.codeFirstHello.sayHi]
+    - servicecomb-0.x-scenario: [/servicecomb/case, /CodeFirstPojoHelloProviderImpl/sayHi,
+                                codefirst.codeFirstSpringmvcHelloClient.say, codefirst.codeFirstHello.sayHi]
   heartbeat: []
 segmentItems:
-  - applicationCode: servicecomb-provider
-    segmentSize: gt 1
+  - applicationCode: servicecomb-0.x-scenario
+    segmentSize: gt 2
     segments:
       - segmentId: not null
         spans:
@@ -47,51 +45,11 @@ segmentItems:
             tags:
               - {key: url, value: /sayHi}
             refs:
-              - {parentEndpointId: 0, parentEndpoint: codefirstClient.codeFirstSpringmvcHello.say,
+              - {parentEndpointId: 0, parentEndpoint: codefirst.codeFirstSpringmvcHelloClient.say,
                  networkAddressId: 0, entryEndpointId: 0, refType: CrossProcess, parentSpanId: 1,
-                 parentTraceSegmentId: not null, parentServiceInstanceId: 2,
-                 networkAddress: not null, entryEndpoint: codefirstClient.codeFirstSpringmvcHello.say,
-                 entryServiceInstanceId: 2}
-  - applicationCode: servicecomb-consumer
-    segmentSize: gt 3
-    segments:
-      - segmentId: not null
-        spans:
-          - operationName: /CodeFirstPojoHelloImpl/sayHi
-            operationId: 0
-            parentSpanId: -1
-            spanId: 0
-            spanLayer: Http
-            startTime: nq 0
-            endTime: nq 0
-            componentId: 59
-            componentName: ''
-            isError: false
-            spanType: Exit
-            peer: not null
-            peerId: 0
-            tags:
-              - {key: http.method, value: POST}
-              - {key: url, value: /CodeFirstPojoHelloImpl/sayHi}
-      - segmentId: not null
-        spans:
-          - operationName: '#/CodeFirstPojoHelloImpl/sayHi'
-            operationId: 0
-            parentSpanId: -1
-            spanId: 0
-            spanLayer: Http
-            startTime: nq 0
-            endTime: nq 0
-            componentId: 59
-            componentName: ''
-            isError: false
-            spanType: Local
-            peer: ''
-            peerId: 0
-            refs:
-              - {parentEndpointId: -1, parentEndpoint: '', networkAddressId: 0, entryEndpointId: -1,
-                 refType: CrossThread, parentSpanId: 0, parentTraceSegmentId: not null,
-                 parentServiceInstanceId: 2, networkAddress: '', entryEndpoint: '', entryServiceInstanceId: 2}
+                 parentTraceSegmentId: not null, parentServiceInstanceId: 1,
+                 networkAddress: not null, entryEndpoint: codefirst.codeFirstSpringmvcHelloClient.say,
+                 entryServiceInstanceId: 1}
       - segmentId: not null
         spans:
           - operationName: codefirst.codeFirstHello.sayHi
@@ -109,7 +67,7 @@ segmentItems:
             peerId: 0
             tags:
               - {key: url, value: /sayHi}
-          - operationName: codefirstClient.codeFirstSpringmvcHello.say
+          - operationName: codefirst.codeFirstSpringmvcHelloClient.say
             operationId: 0
             parentSpanId: -1
             spanId: 0
