+++ test/plugin/scenarios/servicecomb-0.x-scenario/pom.xml
@@ -23,13 +23,6 @@
     <groupId>org.apache.skywalking.apm.testcase</groupId>
     <artifactId>servicecomb-0.x-scenario</artifactId>
     <version>1.0.0</version>
-    <modules>
-        <module>servicecomb-common-schema</module>
-        <module>servicecomb-provider</module>
-        <module>servicecomb-consumer</module>
-        <module>servicecomb-dist</module>
-    </modules>
-    <packaging>pom</packaging>
 
     <modelVersion>4.0.0</modelVersion>
 
@@ -51,13 +44,38 @@
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
-            <dependency>
-                <groupId>org.apache.skywalking.apm.testcase</groupId>
-                <artifactId>servicecomb-common-schema</artifactId>
-                <version>${project.version}</version>
-            </dependency>
         </dependencies>
     </dependencyManagement>
+
+
+    <dependencies>
+        <dependency>
+            <groupId>io.servicecomb</groupId>
+            <artifactId>provider-pojo</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.servicecomb</groupId>
+            <artifactId>transport-highway</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.servicecomb</groupId>
+            <artifactId>transport-rest-vertx</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <version>1.7.7</version>
+        </dependency>
+
+        <dependency>
+            <groupId>io.servicecomb</groupId>
+            <artifactId>provider-springmvc</artifactId>
+            <version>${test.framework.version}</version>
+        </dependency>
+    </dependencies>
 
     <build>
         <finalName>servicecomb-0.x-scenario</finalName>
@@ -69,6 +87,37 @@
                     <target>${compiler.version}</target>
                     <encoding>${project.build.sourceEncoding}</encoding>
                 </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <version>1.5.9.RELEASE</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>repackage</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-assembly-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>assemble</id>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>single</goal>
+                        </goals>
+                        <configuration>
+                            <descriptors>
+                                <descriptor>src/main/assembly/assembly.xml</descriptor>
+                            </descriptors>
+                            <outputDirectory>./target/</outputDirectory>
+                        </configuration>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
