--- test/plugin/scenarios/servicecomb-0.x-scenario/servicecomb-common-schema/src/main/java/org/apache/skywalking/apm/testcase/servicecomb/schemma/Hello.java
@@ -1,29 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.apache.skywalking.apm.testcase.servicecomb.schemma;
-
-
-import org.apache.skywalking.apm.testcase.servicecomb.schemma.models.Person;
-
-public interface Hello {
-
-    String sayHi(String name);
-
-    String sayHello(Person person);
-}
