+++ test/plugin/scenarios/servicecomb-0.x-scenario/src/main/resources/config/log4j.demo.properties
